Module Protocol.Level_repr

type t = private {

level : Tezos_raw_protocol_002_PsYLVpVv.Raw_level_repr.t;
The level of the block relative to genesis. This is also the Shell's notion of level.

level_position : int32;
The level of the block relative to the block that starts protocol alpha. This is specific to the protocol alpha. Other protocols might or might not include a similar notion.

cycle : Tezos_raw_protocol_002_PsYLVpVv.Cycle_repr.t;
The current cycle's number. Note that cycles are a protocol-specific notion. As a result, the cycle number starts at 0 with the first block of protocol alpha.

cycle_position : int32;
The current level of the block relative to the first block of the current cycle.

voting_period : Tezos_raw_protocol_002_PsYLVpVv.Voting_period_repr.t;

voting_period_position : int32;

expected_commitment : bool;

}

type level = t

val (=) : level -> level -> bool

val (<>) : level -> level -> bool

val (<) : level -> level -> bool

val (<=) : level -> level -> bool

val (>=) : level -> level -> bool

val (>) : level -> level -> bool

val compare : level -> level -> int

val equal : level -> level -> bool

val max : level -> level -> level

val min : level -> level -> level

val encoding : level Data_encoding.t

val pp : Stdlib.Format.formatter -> level -> unit

val pp_full : Stdlib.Format.formatter -> level -> unit

val root : Tezos_raw_protocol_002_PsYLVpVv.Raw_level_repr.t -> level

val from_raw : first_level:Tezos_raw_protocol_002_PsYLVpVv.Raw_level_repr.t -> blocks_per_cycle:int32 -> blocks_per_voting_period:int32 -> blocks_per_commitment:int32 -> Tezos_raw_protocol_002_PsYLVpVv.Raw_level_repr.t -> level

val diff : level -> level -> int32