Module Tezos_raw_protocol_002_PsYLVpVv.Delegate_storage

type balance =

| Contract of Contract_repr.t

| Rewards of Tezos_protocol_environment_002_PsYLVpVv.Signature.Public_key_hash.t * Cycle_repr.t

| Fees of Tezos_protocol_environment_002_PsYLVpVv.Signature.Public_key_hash.t * Cycle_repr.t

| Deposits of Tezos_protocol_environment_002_PsYLVpVv.Signature.Public_key_hash.t * Cycle_repr.t

Places where tezzies can be found in the ledger's state.

type balance_update =

| Debited of Tez_repr.t

| Credited of Tez_repr.t

A credit or debit of tezzies to a balance.

type balance_updates = (balance * balance_update) list

A list of balance updates. Duplicates may happen.

val balance_updates_encoding : balance_updates Tezos_protocol_environment_002_PsYLVpVv.Data_encoding.t

val cleanup_balance_updates : balance_updates -> balance_updates

Remove zero-valued balances from a list of updates.

type frozen_balance = {

deposit : Tez_repr.t;

fees : Tez_repr.t;

rewards : Tez_repr.t;

}

val is_delegatable : Raw_context.t -> Contract_repr.t -> bool Tezos_protocol_environment_002_PsYLVpVv.Error_monad.tzresult Tezos_protocol_environment_002_PsYLVpVv.Lwt.t

Is the contract eligible to delegation ?

val init : Raw_context.t -> Contract_repr.t -> Tezos_protocol_environment_002_PsYLVpVv.Signature.Public_key_hash.t -> Raw_context.t Tezos_protocol_environment_002_PsYLVpVv.Error_monad.tzresult Tezos_protocol_environment_002_PsYLVpVv.Lwt.t

Allow to register a delegate when creating an account.

val remove : Raw_context.t -> Contract_repr.t -> Raw_context.t Tezos_protocol_environment_002_PsYLVpVv.Error_monad.tzresult Tezos_protocol_environment_002_PsYLVpVv.Lwt.t

Cleanup delegation when deleting a contract.

val get : Raw_context.t -> Contract_repr.t -> Tezos_protocol_environment_002_PsYLVpVv.Signature.Public_key_hash.t option Tezos_protocol_environment_002_PsYLVpVv.Error_monad.tzresult Tezos_protocol_environment_002_PsYLVpVv.Lwt.t

Reading the current delegate of a contract.

val registered : Raw_context.t -> Tezos_protocol_environment_002_PsYLVpVv.Signature.Public_key_hash.t -> bool Tezos_protocol_environment_002_PsYLVpVv.Lwt.t

val set : Raw_context.t -> Contract_repr.t -> Tezos_protocol_environment_002_PsYLVpVv.Signature.Public_key_hash.t option -> Raw_context.t Tezos_protocol_environment_002_PsYLVpVv.Error_monad.tzresult Tezos_protocol_environment_002_PsYLVpVv.Lwt.t

Updating the delegate of a contract.

When calling this function on an "implicit contract" this function fails, unless when the registered delegate is the contract manager. In the that case, the manager is now registered as a delegate. One cannot unregister a delegate for now. The associate contract is now 'undeletable'.

val set_from_script : Raw_context.t -> Contract_repr.t -> Tezos_protocol_environment_002_PsYLVpVv.Signature.Public_key_hash.t option -> Raw_context.t Tezos_protocol_environment_002_PsYLVpVv.Error_monad.tzresult Tezos_protocol_environment_002_PsYLVpVv.Lwt.t

Same as set ignoring the delegatable flag.

type Tezos_protocol_environment_002_PsYLVpVv.Error_monad.error +=

| Non_delegatable_contract of Contract_repr.contract

| No_deletion of Tezos_protocol_environment_002_PsYLVpVv.Signature.Public_key_hash.t

| Active_delegate

| Current_delegate

| Empty_delegate_account of Tezos_protocol_environment_002_PsYLVpVv.Signature.Public_key_hash.t

| Balance_too_low_for_deposit of {

delegate : Tezos_protocol_environment_002_PsYLVpVv.Signature.Public_key_hash.t;

deposit : Tez_repr.t;

balance : Tez_repr.t;

}

val fold : Raw_context.t -> init:'a -> f: (Tezos_protocol_environment_002_PsYLVpVv.Signature.Public_key_hash.t -> 'a -> 'a Tezos_protocol_environment_002_PsYLVpVv.Lwt.t) -> 'a Tezos_protocol_environment_002_PsYLVpVv.Lwt.t

Iterate on all registered delegates.

val list : Raw_context.t -> Tezos_protocol_environment_002_PsYLVpVv.Signature.Public_key_hash.t list Tezos_protocol_environment_002_PsYLVpVv.Lwt.t

List all registered delegates.

val freeze_deposit : Raw_context.t -> Tezos_protocol_environment_002_PsYLVpVv.Signature.Public_key_hash.t -> Tez_repr.t -> Raw_context.t Tezos_protocol_environment_002_PsYLVpVv.Error_monad.tzresult Tezos_protocol_environment_002_PsYLVpVv.Lwt.t

Various functions to 'freeze' tokens. A frozen 'deposit' keeps its associated rolls. When frozen, 'fees' may trigger new rolls allocation. Rewards won't trigger new rolls allocation until unfrozen.

val freeze_fees : Raw_context.t -> Tezos_protocol_environment_002_PsYLVpVv.Signature.Public_key_hash.t -> Tez_repr.t -> Raw_context.t Tezos_protocol_environment_002_PsYLVpVv.Error_monad.tzresult Tezos_protocol_environment_002_PsYLVpVv.Lwt.t

val freeze_rewards : Raw_context.t -> Tezos_protocol_environment_002_PsYLVpVv.Signature.Public_key_hash.t -> Tez_repr.t -> Raw_context.t Tezos_protocol_environment_002_PsYLVpVv.Error_monad.tzresult Tezos_protocol_environment_002_PsYLVpVv.Lwt.t

val cycle_end : Raw_context.t -> Cycle_repr.t -> Nonce_storage.unrevealed list -> (Raw_context.t * balance_updates * Tezos_protocol_environment_002_PsYLVpVv.Signature.Public_key_hash.t list) Tezos_protocol_environment_002_PsYLVpVv.Error_monad.tzresult Tezos_protocol_environment_002_PsYLVpVv.Lwt.t

Trigger the context maintenance at the end of cycle 'n', i.e.: unfreeze deposit/fees/rewards from 'n - preserved_cycle' ; punish the provided unrevealed seeds (tipically seed from cycle 'n - 1'). Returns a list of account with the amount that was unfrozen for each and the list of deactivated delegates.

val punish : Raw_context.t -> Tezos_protocol_environment_002_PsYLVpVv.Signature.Public_key_hash.t -> Cycle_repr.t -> (Raw_context.t * frozen_balance) Tezos_protocol_environment_002_PsYLVpVv.Error_monad.tzresult Tezos_protocol_environment_002_PsYLVpVv.Lwt.t

Burn all then frozen deposit/fees/rewards for a delegate at a given cycle. Returns the burned amounts.

val has_frozen_balance : Raw_context.t -> Tezos_protocol_environment_002_PsYLVpVv.Signature.Public_key_hash.t -> Cycle_repr.t -> bool Tezos_protocol_environment_002_PsYLVpVv.Error_monad.tzresult Tezos_protocol_environment_002_PsYLVpVv.Lwt.t

Has the given key some frozen tokens in its implicit contract?

val frozen_balance : Raw_context.t -> Tezos_protocol_environment_002_PsYLVpVv.Signature.Public_key_hash.t -> Tez_repr.t Tezos_protocol_environment_002_PsYLVpVv.Error_monad.tzresult Tezos_protocol_environment_002_PsYLVpVv.Lwt.t

Returns the amount of frozen deposit, fees and rewards associated to a given delegate.

val frozen_balance_encoding : frozen_balance Tezos_protocol_environment_002_PsYLVpVv.Data_encoding.t

val frozen_balance_by_cycle_encoding : frozen_balance Cycle_repr.Map.t Tezos_protocol_environment_002_PsYLVpVv.Data_encoding.t

val frozen_balance_by_cycle : Raw_context.t -> Tezos_protocol_environment_002_PsYLVpVv.Signature.Public_key_hash.t -> frozen_balance Cycle_repr.Map.t Tezos_protocol_environment_002_PsYLVpVv.Lwt.t

Returns the amount of frozen deposit, fees and rewards associated to a given delegate, indexed by the cycle by which at the end the balance will be unfrozen.

val full_balance : Raw_context.t -> Tezos_protocol_environment_002_PsYLVpVv.Signature.Public_key_hash.t -> Tez_repr.t Tezos_protocol_environment_002_PsYLVpVv.Error_monad.tzresult Tezos_protocol_environment_002_PsYLVpVv.Lwt.t

Returns the full 'balance' of the implicit contract associated to a given key, i.e. the sum of the spendable balance and of the frozen balance.

val staking_balance : Raw_context.t -> Tezos_protocol_environment_002_PsYLVpVv.Signature.Public_key_hash.t -> Tez_repr.t Tezos_protocol_environment_002_PsYLVpVv.Error_monad.tzresult Tezos_protocol_environment_002_PsYLVpVv.Lwt.t

val delegated_contracts : Raw_context.t -> Tezos_protocol_environment_002_PsYLVpVv.Signature.Public_key_hash.t -> Contract_hash.t list Tezos_protocol_environment_002_PsYLVpVv.Lwt.t

Returns the list of contract that delegated towards a given delegate

val delegated_balance : Raw_context.t -> Tezos_protocol_environment_002_PsYLVpVv.Signature.Public_key_hash.t -> Tez_repr.t Tezos_protocol_environment_002_PsYLVpVv.Error_monad.tzresult Tezos_protocol_environment_002_PsYLVpVv.Lwt.t

val deactivated : Raw_context.t -> Tezos_protocol_environment_002_PsYLVpVv.Signature.Public_key_hash.t -> bool Tezos_protocol_environment_002_PsYLVpVv.Lwt.t

val grace_period : Raw_context.t -> Tezos_protocol_environment_002_PsYLVpVv.Signature.Public_key_hash.t -> Cycle_repr.t Tezos_protocol_environment_002_PsYLVpVv.Error_monad.tzresult Tezos_protocol_environment_002_PsYLVpVv.Lwt.t