Module Tezos_raw_protocol_002_PsYLVpVv__Seed_repr

Tezos Protocol Implementation - Random number generation

This is not expected to be a good cryptographic random number generator. In particular this is supposed to be used in situations where the seed is a globaly known information.

The only expected property is: It should be difficult to find a seed such that the generated sequence is a given one.